A ver si asi se mantiene establa, aun cuando no se ejecuta en una thread aparte.....

+++ include/boost/logging/format/destination/rolling_file.hpp
@@ -128,7 +128,7 @@ namespace detail {
             namespace fs = boost::filesystem;
             m_out = boost::shared_ptr< std::basic_ofstream<char_type> >(new std::basic_ofstream<char_type>( file_name(m_cur_idx).c_str(),
                 m_flags.extra_flags() | std::ios_base::out | std::ios_base::app));
-            if ( fs::file_size( file_name(m_cur_idx)) > m_flags.max_size_bytes()) {
+            if ( static_cast<int>( fs::file_size( file_name(m_cur_idx))) > m_flags.max_size_bytes()) {
                 // this file is already full - clear it first
                 m_out = boost::shared_ptr< std::basic_ofstream<char_type> >(new std::basic_ofstream<char_type>( file_name(m_cur_idx).c_str(),
                     m_flags.extra_flags() | std::ios_base::out | std::ios_base::trunc));
